Add extractor worker!
