+++ package.json
@@ -16,6 +16,7 @@
     "collector-server": "NODE_PATH=dist node --optimize_for_size --max_old_space_size=460 --gc_interval=100 dist/collector/server.js",
     "collector-worker": "NODE_PATH=dist node --optimize_for_size --max_old_space_size=460 --gc_interval=100 dist/collector/worker.js",
     "extractor-server": "NODE_PATH=dist node --optimize_for_size --max_old_space_size=460 --gc_interval=100 dist/extractor/server.js",
+    "extractor-worker": "NODE_PATH=dist node --optimize_for_size --max_old_space_size=460 --gc_interval=100 dist/extractor/worker.js",
     "eyeshade-server": "NODE_PATH=dist node --optimize_for_size --max_old_space_size=460 --gc_interval=100 dist/eyeshade/server.js",
     "eyeshade-worker": "NODE_PATH=dist node --optimize_for_size --max_old_space_size=460 --gc_interval=100 dist/eyeshade/worker.js",
     "helper-server": "NODE_PATH=dist node --optimize_for_size --max_old_space_size=460 --gc_interval=100 dist/helper/server.js",
@@ -23,11 +24,12 @@
     "ledger-server": "NODE_PATH=dist node --optimize_for_size --max_old_space_size=460 --gc_interval=100 dist/ledger/server.js",
     "ledger-worker": "NODE_PATH=dist node --optimize_for_size --max_old_space_size=460 --gc_interval=100 dist/ledger/worker.js",
     "lint": "standard",
-    "start": "echo use 'npm run start-[balance|collector|extractor|eyeshade|helper|ledger]' or 'npm run start-[collector|eyeshade|ledger]-worker'",
+    "start": "echo use 'npm run start-[balance|collector|extractor|eyeshade|helper|ledger]' or 'npm run start-[collector|extractor|eyeshade|ledger]-worker'",
     "start-balance": "NODE_PATH=. DEBUG=*,-babel,-mongo:*,mongo:queries SERVICE=balance ./node_modules/.bin/babel-node balance/server.js",
     "start-collector": "NODE_PATH=. DEBUG=*,-babel,-mongo:*,mongo:queries SERVICE=collector ./node_modules/.bin/babel-node collector/server.js",
     "start-collector-worker": "NODE_PATH=. DEBUG=*,-babel,-mongo:*,mongo:queries SERVICE=collector ./node_modules/.bin/babel-node --optimize_for_size --max_old_space_size=460 --gc_interval=100 collector/worker.js",
     "start-extractor": "NODE_PATH=. DEBUG=*,-babel,-mongo:*,mongo:queries SERVICE=extractor ./node_modules/.bin/babel-node extractor/server.js",
+    "start-extractor-worker": "NODE_PATH=. DEBUG=*,-babel,-mongo:*,mongo:queries SERVICE=extractor ./node_modules/.bin/babel-node --optimize_for_size --max_old_space_size=460 --gc_interval=100 extractor/worker.js",
     "start-eyeshade": "NODE_PATH=. DEBUG=*,-babel,-mongo:*,mongo:queries SERVICE=eyeshade ./node_modules/.bin/babel-node eyeshade/server.js",
     "start-eyeshade-worker": "NODE_PATH=. DEBUG=*,-babel,-mongo:*,mongo:queries SERVICE=eyeshade ./node_modules/.bin/babel-node --optimize_for_size --max_old_space_size=460 --gc_interval=100 eyeshade/worker.js",
     "start-helper": "NODE_PATH=. DEBUG=*,-babel,-mongo:*,mongo:queries SERVICE=helper ./node_modules/.bin/babel-node helper/server.js",
